@@ -1,7 +1,7 @@
 {
   "name": "vue-table-sorter",
-  "version": "1.0.0",
-  "description": "Vue.js 2 component for table sort.",
+  "version": "2.0.0",
+  "description": "Vue.js component for table sort.",
   "main": "dist/vue-table-sorter.js",
   "browser": "dist/vue-table-sorter.min.js",
   "files": [
@@ -36,8 +36,9 @@
   "devDependencies": {
     "@babel/core": "^7.19.3",
     "@babel/preset-env": "^7.19.3",
-    "@vue/test-utils": "^1.1.4",
-    "@vue/vue2-jest": "^29.1.1",
+    "@vue/compiler-sfc": "^3.2.40",
+    "@vue/test-utils": "^2",
+    "@vue/vue3-jest": "^29.1.1",
     "babel-jest": "^29.1.2",
     "babel-loader": "^8.2.5",
     "bootstrap": "^4.6.0",
@@ -49,17 +50,16 @@
     "jest-environment-jsdom": "^29.1.2",
     "mini-css-extract-plugin": "^2.6.1",
     "style-loader": "^3.3.1",
-    "vue": "^2.7",
-    "vue-loader": "^15.10",
-    "vue-template-compiler": "^2.7",
+    "vue": "^3",
+    "vue-loader": "^17",
     "webpack": "^5.74.0",
     "webpack-cli": "^4.10.0",
     "webpack-dev-server": "^4.11.1"
   },
   "peerDependencies": {
-    "vue": "^2.7.0"
+    "vue": "^3.2.0"
   },
   "engines": {
-    "node": ">=12.13.0"
+    "node": ">=14.20.0"
   }
 }
