@@ -1,12 +1,8 @@
 'use strict';
 
-import Vue from 'vue';
-
+import {createApp} from 'vue';
 import 'bootstrap/dist/css/bootstrap.min.css';
 
 import App from './App.vue';
 
-new Vue({
-  el: '#app',
-  render: h => h(App),
-});
+createApp(App).mount('#app')
