@@ -13,9 +13,8 @@ module.exports = {
       path.resolve(__dirname, 'node_modules'),
     ],
     alias: {
-      'vue$': 'vue/dist/vue.runtime.esm.js'
+      vue: "@vue/runtime-dom"
     },
-    extensions: ['.js', '.json', '.vue']
   },
   entry: {
     'vue-table-sorter': './src/index.js',
